Extract analysis logic
